backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
djangorestframework==3.16.1
deep-translator==1.11.4
idna==3.11
insightface==0.7.3
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
pillow==12.1.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
